Various Updates (see desc)
While trying to make the enqueue method more efficient by sorting the left from the right nodes in the tree I rediscovered the infinite loop that wasn't present on other machine but was on mine. Because of this I needed to keep the sorting of left and right nodes in the freq table. Also now functioning building of tree.