Important Comment Added (see desc)
A lot of confusion was caused due to IntelliJ displaying the tree in the console in reverse. As a result, a large portion of time was dedicated to why the code was inverting the boolean lists. However, this was eventually solved and gave me a better understanding of the code and platform.